+++ crates/l2/proposer/l1_watcher.rs
@@ -4,7 +4,7 @@ use crate::{
 };
 use bytes::Bytes;
 use ethereum_types::{Address, BigEndianHash, H256, U256};
-use ethrex_blockchain::{constants::TX_GAS_COST, mempool};
+use ethrex_blockchain::{constants::TX_GAS_COST, Blockchain};
 use ethrex_common::types::{Signable, Transaction};
 use ethrex_rpc::clients::eth::{errors::EthClientError, eth_sender::Overrides, EthClient};
 use ethrex_rpc::types::receipt::RpcLog;
@@ -15,11 +15,11 @@ use std::{cmp::min, ops::Mul, time::Duration};
 use tokio::time::sleep;
 use tracing::{debug, error, info, warn};
 
-pub async fn start_l1_watcher(store: Store) -> Result<(), ConfigError> {
+pub async fn start_l1_watcher(store: Store, blockchain: Blockchain) -> Result<(), ConfigError> {
     let eth_config = EthConfig::from_env()?;
     let watcher_config = L1WatcherConfig::from_env()?;
     let mut l1_watcher = L1Watcher::new_from_config(watcher_config, eth_config).await?;
-    l1_watcher.run(&store).await;
+    l1_watcher.run(&store, &blockchain).await;
     Ok(())
 }
 
@@ -55,9 +55,9 @@ impl L1Watcher {
         })
     }
 
-    pub async fn run(&mut self, store: &Store) {
+    pub async fn run(&mut self, store: &Store, blockchain: &Blockchain) {
         loop {
-            if let Err(err) = self.main_logic(store).await {
+            if let Err(err) = self.main_logic(store, blockchain).await {
                 error!("L1 Watcher Error: {}", err);
             }
 
@@ -65,7 +65,11 @@ impl L1Watcher {
         }
     }
 
-    async fn main_logic(&mut self, store: &Store) -> Result<(), L1WatcherError> {
+    async fn main_logic(
+        &mut self,
+        store: &Store,
+        blockchain: &Blockchain,
+    ) -> Result<(), L1WatcherError> {
         loop {
             sleep(self.check_interval).await;
 
@@ -78,7 +82,7 @@ impl L1Watcher {
 
             let pending_deposit_logs = self.get_pending_deposit_logs().await?;
             let _deposit_txs = self
-                .process_logs(logs, &pending_deposit_logs, store)
+                .process_logs(logs, &pending_deposit_logs, store, blockchain)
                 .await?;
         }
     }
@@ -162,6 +166,7 @@ impl L1Watcher {
         logs: Vec<RpcLog>,
         pending_deposit_logs: &[H256],
         store: &Store,
+        blockchain: &Blockchain,
     ) -> Result<Vec<H256>, L1WatcherError> {
         let mut deposit_txs = Vec::new();
 
@@ -267,10 +272,9 @@ impl L1Watcher {
                 .await?;
             mint_transaction.sign_inplace(&self.l2_proposer_pk);
 
-            match mempool::add_transaction(
-                Transaction::PrivilegedL2Transaction(mint_transaction),
-                store,
-            ) {
+            match blockchain
+                .add_transaction_to_pool(Transaction::PrivilegedL2Transaction(mint_transaction))
+            {
                 Ok(hash) => {
                     info!("Mint transaction added to mempool {hash:#x}",);
                     deposit_txs.push(hash);
