+++ crates/l2/proposer/mod.rs
@@ -3,6 +3,7 @@ use std::time::Duration;
 use crate::utils::config::{errors::ConfigError, proposer::ProposerConfig, read_env_file};
 use errors::ProposerError;
 use ethereum_types::Address;
+use ethrex_blockchain::Blockchain;
 use ethrex_rpc::clients::EngineApiConfig;
 use ethrex_storage::Store;
 use tokio::task::JoinSet;
@@ -25,7 +26,7 @@ pub struct Proposer {
     jwt_secret: Vec<u8>,
 }
 
-pub async fn start_proposer(store: Store) {
+pub async fn start_proposer(store: Store, blockchain: Blockchain) {
     info!("Starting Proposer");
 
     if let Err(e) = read_env_file() {
@@ -34,7 +35,10 @@ pub async fn start_proposer(store: Store) {
     }
 
     let mut task_set = JoinSet::new();
-    task_set.spawn(l1_watcher::start_l1_watcher(store.clone()));
+    task_set.spawn(l1_watcher::start_l1_watcher(
+        store.clone(),
+        blockchain.clone(),
+    ));
     task_set.spawn(l1_committer::start_l1_committer(store.clone()));
     task_set.spawn(prover_server::start_prover_server(store.clone()));
     task_set.spawn(start_proposer_server(store.clone()));
