+++ crates/networking/p2p/network.rs
@@ -11,6 +11,7 @@ use crate::{
     },
     rlpx::utils::log_peer_error,
 };
+use ethrex_blockchain::Blockchain;
 use ethrex_common::H512;
 use ethrex_storage::Store;
 use k256::{
@@ -47,6 +48,7 @@ pub struct P2PContext {
     pub signer: SigningKey,
     pub table: Arc<Mutex<KademliaTable>>,
     pub storage: Store,
+    pub blockchain: Blockchain,
     pub(crate) broadcast: RLPxConnBroadcastSender,
     pub local_node: Node,
     pub enr_seq: u64,
@@ -59,6 +61,7 @@ pub async fn start_network(
     signer: SigningKey,
     peer_table: Arc<Mutex<KademliaTable>>,
     storage: Store,
+    blockchain: Blockchain,
 ) -> Result<(), NetworkError> {
     let (channel_broadcast_send_end, _) = tokio::sync::broadcast::channel::<(
         tokio::task::Id,
@@ -75,6 +78,7 @@ pub async fn start_network(
         signer,
         table: peer_table,
         storage,
+        blockchain,
         broadcast: channel_broadcast_send_end,
     };
     let discovery = Discv4Server::try_new(context.clone())
