+++ crates/networking/p2p/rlpx/eth/transactions.rs
@@ -1,7 +1,7 @@
 use bytes::BufMut;
 use bytes::Bytes;
 use ethrex_blockchain::error::MempoolError;
-use ethrex_blockchain::mempool;
+use ethrex_blockchain::Blockchain;
 use ethrex_common::types::BlobsBundle;
 use ethrex_common::types::P2PTransaction;
 use ethrex_common::types::WrappedEIP4844Transaction;
@@ -10,7 +10,7 @@ use ethrex_rlp::{
     error::{RLPDecodeError, RLPEncodeError},
     structs::{Decoder, Encoder},
 };
-use ethrex_storage::{error::StoreError, Store};
+use ethrex_storage::error::StoreError;
 
 use crate::rlpx::utils::log_peer_warn;
 use crate::rlpx::{
@@ -72,7 +72,10 @@ pub(crate) struct NewPooledTransactionHashes {
 }
 
 impl NewPooledTransactionHashes {
-    pub fn new(transactions: Vec<Transaction>, storage: &Store) -> Result<Self, StoreError> {
+    pub fn new(
+        transactions: Vec<Transaction>,
+        blockchain: &Blockchain,
+    ) -> Result<Self, StoreError> {
         let transactions_len = transactions.len();
         let mut transaction_types = Vec::with_capacity(transactions_len);
         let mut transaction_sizes = Vec::with_capacity(transactions_len);
@@ -89,8 +92,9 @@ impl NewPooledTransactionHashes {
                 // Network representation for PooledTransactions
                 // https://eips.ethereum.org/EIPS/eip-4844#networking
                 Transaction::EIP4844Transaction(eip4844_tx) => {
-                    let tx_blobs_bundle = storage
-                        .get_blobs_bundle_from_pool(transaction_hash)?
+                    let tx_blobs_bundle = blockchain
+                        .mempool
+                        .get_blobs_bundle(transaction_hash)?
                         .unwrap_or(BlobsBundle::empty());
                     eip4844_tx.rlp_length_as_pooled_tx(&tx_blobs_bundle)
                 }
@@ -105,8 +109,13 @@ impl NewPooledTransactionHashes {
         })
     }
 
-    pub fn get_transactions_to_request(&self, storage: &Store) -> Result<Vec<H256>, StoreError> {
-        storage.filter_unknown_transactions(&self.transaction_hashes)
+    pub fn get_transactions_to_request(
+        &self,
+        blockchain: &Blockchain,
+    ) -> Result<Vec<H256>, StoreError> {
+        blockchain
+            .mempool
+            .filter_unknown_transactions(&self.transaction_hashes)
     }
 }
 
@@ -166,12 +175,12 @@ impl GetPooledTransactions {
         }
     }
 
-    pub fn handle(&self, store: &Store) -> Result<PooledTransactions, StoreError> {
+    pub fn handle(&self, blockchain: &Blockchain) -> Result<PooledTransactions, StoreError> {
         // TODO(#1615): get transactions in batch instead of iterating over them.
         let txs = self
             .transaction_hashes
             .iter()
-            .map(|hash| Self::get_p2p_transaction(hash, store))
+            .map(|hash| Self::get_p2p_transaction(hash, blockchain))
             // Return an error in case anything failed.
             .collect::<Result<Vec<_>, _>>()?
             .into_iter()
@@ -189,9 +198,9 @@ impl GetPooledTransactions {
     /// Gets a p2p transaction given a hash.
     fn get_p2p_transaction(
         hash: &H256,
-        store: &Store,
+        blockchain: &Blockchain,
     ) -> Result<Option<P2PTransaction>, StoreError> {
-        let Some(tx) = store.get_transaction_by_hash_from_pool(*hash)? else {
+        let Some(tx) = blockchain.mempool.get_transaction_by_hash(*hash)? else {
             return Ok(None);
         };
         let result = match tx {
@@ -199,7 +208,7 @@ impl GetPooledTransactions {
             Transaction::EIP2930Transaction(itx) => P2PTransaction::EIP2930Transaction(itx),
             Transaction::EIP1559Transaction(itx) => P2PTransaction::EIP1559Transaction(itx),
             Transaction::EIP4844Transaction(itx) => {
-                let Some(bundle) = store.get_blobs_bundle_from_pool(*hash)? else {
+                let Some(bundle) = blockchain.mempool.get_blobs_bundle(*hash)? else {
                     return Err(StoreError::Custom(format!(
                         "Blob transaction present without its bundle: hash {}",
                         hash
@@ -263,10 +272,10 @@ impl PooledTransactions {
 
     /// Saves every incoming pooled transaction to the mempool.
 
-    pub fn handle(self, node: &Node, store: &Store) -> Result<(), MempoolError> {
+    pub fn handle(self, node: &Node, blockchain: &Blockchain) -> Result<(), MempoolError> {
         for tx in self.pooled_transactions {
             if let P2PTransaction::EIP4844TransactionWithBlobs(itx) = tx {
-                if let Err(e) = mempool::add_blob_transaction(itx.tx, itx.blobs_bundle, store) {
+                if let Err(e) = blockchain.add_blob_transaction_to_pool(itx.tx, itx.blobs_bundle) {
                     log_peer_warn(node, &format!("Error adding transaction: {}", e));
                     continue;
                 }
@@ -274,7 +283,7 @@ impl PooledTransactions {
                 let regular_tx = tx
                     .try_into()
                     .map_err(|error| MempoolError::StoreError(StoreError::Custom(error)))?;
-                if let Err(e) = mempool::add_transaction(regular_tx, store) {
+                if let Err(e) = blockchain.add_transaction_to_pool(regular_tx) {
                     log_peer_warn(node, &format!("Error adding transaction: {}", e));
                     continue;
                 }
