+++ crates/networking/rpc/engine/fork_choice.rs
@@ -242,10 +242,12 @@ fn handle_forkchoice(
             // Remove included transactions from the mempool after we accept the fork choice
             // TODO(#797): The remove of transactions from the mempool could be incomplete (i.e. REORGS)
             if let Ok(Some(block)) = context.storage.get_block_by_hash(head.compute_block_hash()) {
-                context
-                    .storage
-                    .remove_transactions_from_pool(&block.body.transactions)
-                    .map_err(|err| RpcErr::Internal(err.to_string()))?;
+                for tx in &block.body.transactions {
+                    context
+                        .blockchain
+                        .remove_transaction_from_pool(&tx.compute_hash())
+                        .map_err(|err| RpcErr::Internal(err.to_string()))?;
+                }
             } else {
                 return Err(RpcErr::Internal(
                     "Failed to get block by hash to remove transactions from the mempool"
